Optimized minHeight and maxHeight calculation Optimized initialization of default attributes Vars that should be floats now float; vars that should be doubles now double.

+++ class/HPGrowingTextView.h
@@ -52,8 +52,8 @@
 @interface HPGrowingTextView : UIView <UITextViewDelegate> {
 	HPTextViewInternal *internalTextView;	
 	
-	int minHeight;
-	int maxHeight;
+	CGFloat minHeight;
+	CGFloat maxHeight;
 	
 	//class properties
 	int maxNumberOfLines;
